+++ C/include/array_g.h
@@ -1,5 +1,7 @@
 #pragma once
 
-void* at_g(void*, unsigned int pos, unsigned int size);
-void cp_g(void*, const void*, unsigned int size);
-void swap_g(void*, void*, unsigned int size);
+#include <stdlib.h>
+
+void* at_g(void*, size_t pos, size_t size);
+void cp_g(void*, const void*, size_t size);
+void swap_g(void*, void*, size_t size);
