+++ C/include/dictionary.h
@@ -32,7 +32,7 @@ int Dictionary_get(Dictionary dictionary, const void* key, void** result);
 void Dictionary_delete(Dictionary dictionary, const void* key);
 
 // Returns the number of distinct keys stores in the dictionary.
-unsigned int Dictionary_size(Dictionary dictionary);
+size_t Dictionary_size(Dictionary dictionary);
 int Dictionary_empty(Dictionary dictionary);
 
 // Returns an evaluation of the efficiency of the dictionary. Its actual
