+++ C/include/insertion_sort.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <stdlib.h>
+
+
 typedef int (*ISCompareFun)(const void*, const void*);
 
-void insertion_sort(void** array, unsigned int count, ISCompareFun compare);
+void insertion_sort(void** array, size_t count, ISCompareFun compare);
