+++ C/include/key_info.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdlib.h>
+
 /*
  * KeyInfo maintains stores a pair of functions to compare and
  * create hashes of given data.
@@ -7,7 +9,7 @@
 
 typedef struct _KeyInfo* KeyInfo;
 typedef int (*KIComparator)(const void*, const void*);
-typedef unsigned int (*KIHash)(const void*);
+typedef size_t (*KIHash)(const void*);
 
 /* Constructor and destructor*/
 KeyInfo KeyInfo_new( KIComparator, KIHash );
@@ -25,6 +27,6 @@ KIHash KeyInfo_hash(KeyInfo);
 int KeyInfo_string_compare(const void* e1, const void* e2);
 int KeyInfo_int_compare(const void* e1, const void* e2);
 int KeyInfo_double_compare(const void* e1, const void* e2);
-unsigned int KeyInfo_string_hash(const void* e1);
-unsigned int KeyInfo_int_hash(const void* e);
-unsigned int KeyInfo_double_hash(const void* e);
+size_t KeyInfo_string_hash(const void* e1);
+size_t KeyInfo_int_hash(const void* e);
+size_t KeyInfo_double_hash(const void* e);
