+++ C/include/stack.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <stdlib.h>
+
 typedef struct _Stack* Stack;
 
 
-Stack Stack_new(unsigned int capacity);
+Stack Stack_new(size_t capacity);
 void Stack_free(Stack stack);
 
 void* Stack_top(Stack stack);
