+++ C/src/array_g.c
@@ -9,16 +9,16 @@ typedef unsigned char uchar;
 static uchar SWAP_BUFFER[SWAP_BUFFER_SIZE];
 
 
-void* at_g(void* array, unsigned int pos, unsigned int size) {
+void* at_g(void* array, size_t pos, size_t size) {
   uchar* carray = (uchar*) array;
   return (void*) &carray[pos*size];
 }
 
-void cp_g(void* dst, const void* src, unsigned int size) {
+void cp_g(void* dst, const void* src, size_t size) {
     memcpy(dst, src, size);
 }
 
-void swap_g(void* el1, void* el2, unsigned int size) {
+void swap_g(void* el1, void* el2, size_t size) {
   assert(size <= SWAP_BUFFER_SIZE);
 
   cp_g(SWAP_BUFFER, el1, size);
