+++ C/src/insertion_sort.c
@@ -1,8 +1,8 @@
 #include "insertion_sort.h"
 #include <stdio.h>
 
-static void insert(void** array, unsigned int elem_pos, ISCompareFun compare) {
-  unsigned int i=elem_pos;
+static void insert(void** array, size_t elem_pos, ISCompareFun compare) {
+  size_t i=elem_pos;
   void* elem = array[elem_pos];
   for(; i>0 && compare(elem, array[i-1]) < 0; --i) {
     array[i] = array[i-1];
@@ -12,8 +12,8 @@ static void insert(void** array, unsigned int elem_pos, ISCompareFun compare) {
 }
 
 
-void insertion_sort(void** array, unsigned int count, ISCompareFun compare) {
-  for(unsigned int i=1; i<count; ++i) {
+void insertion_sort(void** array, size_t count, ISCompareFun compare) {
+  for(size_t i=1; i<count; ++i) {
     insert(array, i, compare);
   }
 }
